feat: create NetworkMonitor
